requirements.txt
click==8.1.3
joblib==1.2.0
nltk==3.8.1
packaging==23.1
Pillow==9.5.0
pypdfium2==4.13.0
PyQt6==6.5.0
PyQt6-Qt6==6.5.1
PyQt6-sip==13.5.1
pytesseract==0.3.10
regex==2023.5.5
tqdm==4.65.0
